requirements.txt
aldjemy==0.10.1
amqp==2.5.2
argon2-cffi==19.1.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
Django==2.2.10
django-coverage-plugin==1.6.0
django-environ==0.4.5
django-extensions==2.1.7
django-model-utils==3.1.2
django-rest-swagger==2.2.0
django-user-agents==0.4.0
djangorestframework==3.10.3
djangorestframework-csv==2.1.0
entrypoints==0.3
flake8==3.7.5
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
openapi-codec==1.3.2
packaging==20.3
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2==2.8
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pyparsing==2.4.6
pytest==5.2.1
pytest-cov==2.8.1
pytest-django==3.5.1
pytest-mock==1.11.1
pytest-sugar==0.9.2
python-dateutil==2.8.0
python-slugify==3.0.4
pytz==2019.1
PyYAML==5.3
redis==3.4.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
SQLAlchemy==1.3.15
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.3
typed-ast==1.4.1
ua-parser==0.10.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.24.3
user-agents==2.1
vine==1.3.0
wcwidth==0.1.8
wrapt==1.11.2
zipp==3.1.0
